Securing APIs module finished.

+++ MyCodeCamp/MyCodeCamp/Controllers/SpeakersController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using MyCodeCamp.Data;
@@ -8,6 +9,7 @@ using MyCodeCamp.Models;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 
 namespace MyCodeCamp.Controllers
 {
@@ -18,12 +20,14 @@ namespace MyCodeCamp.Controllers
         private readonly ILogger<SpeakersController> logger;
         private readonly IMapper mapper;
         private readonly ICampRepository repository;
+        private UserManager<CampUser> userManager;
 
-        public SpeakersController(ILogger<SpeakersController> logger, IMapper mapper, ICampRepository repository)
+        public SpeakersController(ILogger<SpeakersController> logger, IMapper mapper, ICampRepository repository, UserManager<CampUser> userManager)
         {
             this.logger = logger;
             this.mapper = mapper;
             this.repository = repository;
+            this.userManager = userManager;
         }
 
         [HttpGet]
@@ -48,6 +52,7 @@ namespace MyCodeCamp.Controllers
             return Ok(mapper.Map<SpeakerModel>(speaker));
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> Post(string moniker, [FromBody]SpeakerModel model)
         {
@@ -60,12 +65,18 @@ namespace MyCodeCamp.Controllers
                 var speaker = mapper.Map<Speaker>(model);
                 speaker.Camp = camp;
 
-                repository.Add(speaker);
-
-                if (await repository.SaveAllAsync())
+                var campUser = await userManager.FindByNameAsync(User.Identity.Name);
+                if (campUser != null)
                 {
-                    var url = Url.Link("SpeakerGet", new { moniker = camp.Moniker, id = speaker.Id });
-                    return Created(url, mapper.Map<SpeakerModel>(speaker));
+                    speaker.User = campUser;
+
+                    repository.Add(speaker);
+
+                    if (await repository.SaveAllAsync())
+                    {
+                        var url = Url.Link("SpeakerGet", new { moniker = camp.Moniker, id = speaker.Id });
+                        return Created(url, mapper.Map<SpeakerModel>(speaker));
+                    }
                 }
             }
             catch (Exception exception)
@@ -76,6 +87,7 @@ namespace MyCodeCamp.Controllers
             return BadRequest("Could not add new speaker");
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<IActionResult> Put(string moniker, int id, [FromBody] SpeakerModel model)
         {
@@ -88,6 +100,9 @@ namespace MyCodeCamp.Controllers
 
                 if (speaker.Camp.Moniker != moniker)
                     return BadRequest("Speaker not in specified Camp");
+
+                if (speaker.User.UserName != User.Identity.Name)
+                    return Forbid();
 
                 mapper.Map(model, speaker);
 
@@ -102,6 +117,7 @@ namespace MyCodeCamp.Controllers
             return BadRequest("Could not update Speaker");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(string moniker, int id)
         {
@@ -114,6 +130,9 @@ namespace MyCodeCamp.Controllers
 
                 if (speaker.Camp.Moniker != moniker)
                     return BadRequest("Speaker not in specified Camp");
+
+                if (speaker.User.UserName != User.Identity.Name)
+                    return Forbid();
 
                 repository.Delete(speaker);
